Analyzed experience vs subscriber count and distribution of ages
In gio_testing, created a bar plot showing the proportions of players subscribed or not amongst the different experience groups, aswell as created a bar plot showing how many players are in different ranges of ages: less than 10, between 10 and 20, 20 to 30, 30 to 50, and older than 50.

I found that the experience category has no effect on subscriber count, and that by far the most common ages are between 10 to 30, in particular 10 to 20.