Visualizing experience in age groups and hours played
Created proportional bar plots to explore the distribution of experience amongst the different age groups and amongst ranges of hours played. In both instances, I found the data distributions to be mostly consistent, further suggesting the irrelevance of the experience category.

The only notable outlier in either case was the "pro" category.